+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes in 12.2
 - Security: move RosarioSIS version from Login screen to Portal in index.php & Portal.php
 - Add `composer.json` to Files list (required for add-ons) in modules/README.md & plugins/README.md
 - Check subject, course & course period ID are valid for current school & year in Courses.php
+- Fix SQL error: user just deleted current Marking Period, reload Side menu in MarkingPeriods.php
 
 Changes in 12.1.3
 -----------------